requirements.txt
flask>=3.0
torch>=2.0
torchvision>=0.15
timm>=0.9
transformers>=4.30
opencv-python>=4.8
Pillow>=10.0
numpy>=1.24
